jarvis/pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "langfuse>=2.0.0,<3.0.0",
    "litellm>=1.83.13",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.5.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.0.1",
    "structlog>=25.5.0",
    "uvicorn>=0.46.0",
]
